完成登录注册功能

+++ user-service/src/main/java/com/hiiro/service/impl/UserServiceImpl.java
@@ -1,7 +1,6 @@
 package com.hiiro.service.impl;
 
 import cn.hutool.core.bean.BeanUtil;
-import cn.hutool.core.date.DateUtil;
 import com.baomidou.mybatisplus.core.conditions.query.LambdaQueryWrapper;
 import com.baomidou.mybatisplus.core.conditions.update.LambdaUpdateWrapper;
 import com.baomidou.mybatisplus.extension.service.impl.ServiceImpl;
@@ -9,7 +8,6 @@ import com.hiiro.entity.ResultCodeEnum;
 import com.hiiro.entity.ResultData;
 import com.hiiro.entity.User;
 import com.hiiro.entity.dto.UserDTO;
-import com.hiiro.exp.UserException;
 import com.hiiro.mapper.UserMapper;
 import com.hiiro.service.UserService;
 import com.hiiro.utils.MyJwtUtil;
@@ -65,7 +63,7 @@ public class UserServiceImpl extends ServiceImpl<UserMapper, User> implements Us
     public ResultData<String> register(User user) {
         // 验证用户名是否已存在
         if (Objects.nonNull(getUserByUsername(user.getUsername()))) {
-            throw new UserException(ResultCodeEnum.DATABASE_INSERT_ERROR, "用户已存在!");
+            return ResultData.fail(ResultCodeEnum.INTERNAL_SERVER_ERROR, "用户已存在!");
         }
         // 加密用户密码
         user.setPassword(passwordEncoder.encode(user.getPassword()));
@@ -77,12 +75,12 @@ public class UserServiceImpl extends ServiceImpl<UserMapper, User> implements Us
                 user.setNickname("hiri_" + user.getUid());
                 // 更新用户昵称
                 if (updateUserById(user) != 1) {
-                    throw new UserException(ResultCodeEnum.DATABASE_UPDATE_ERROR, "更新用户信息失败");
+                    return ResultData.fail(ResultCodeEnum.INTERNAL_SERVER_ERROR, "更新用户信息失败");
                 }
             }
             return ResultData.success("注册成功，欢迎加入hirihiri！");
         } else {
-            throw new UserException(ResultCodeEnum.DATABASE_INSERT_ERROR, "用户注册失败");
+            return ResultData.fail(ResultCodeEnum.INTERNAL_SERVER_ERROR, "用户注册失败");
         }
     }
 
@@ -96,27 +94,26 @@ public class UserServiceImpl extends ServiceImpl<UserMapper, User> implements Us
     @Override
     public ResultData<HashMap<String, Object>> login(User user) {
         // 创建一个UsernamePasswordAuthenticationToken对象，用于认证用户
-        UsernamePasswordAuthenticationToken authenticationToken = new UsernamePasswordAuthenticationToken(user.getUsername(), user.getPassword());
+        UsernamePasswordAuthenticationToken authenticationToken = new UsernamePasswordAuthenticationToken(
+                user.getUsername(), user.getPassword());
         // 使用authenticationProvider对提供的用户名和密码进行验证
         Authentication authenticate = authenticationProvider.authenticate(authenticationToken);
         // 获取认证通过后的用户详细信息
         UserDetailsImpl loginUser = (UserDetailsImpl) authenticate.getPrincipal();
         if (loginUser.getUser().getState() != 0) {
-            throw new UserException(ResultCodeEnum.USER_BANNED_OR_DELETED, ResultCodeEnum.USER_BANNED_OR_DELETED.getMessage());
+            return ResultData.fail(ResultCodeEnum.UNAUTHORIZED, "用户被封禁或已注销");
         }
-        // 将用户ID（UID）转换为字符串形式，以便放入JWT令牌中
-        String uid = String.valueOf(loginUser.getUser().getUid());
+        // 取出用户uid放入JWT令牌中
+        Long uid = loginUser.getUser().getUid();
         // 创建默认的JWT令牌，其中包含用户的UID作为声明的一部分
         String token = jwtUtil.createDefaultJwtToken(new HashMap<>(Map.of("uid", uid)));
-        // 更新用户登录状态
-        if (updateUserById(loginUser.getUser()) == 1) {
-            // 登陆成功并成功更新登陆状态后将用户信息存入redis
-            redisUtil.setObjectWithExpire("user:" + uid, loginUser.getUser(), TimeUnit.SECONDS);
-            // 返回token和用户信息给前端
-            return ResultData.success(new HashMap<String, Object>(Map.of("user", BeanUtil.copyProperties(getUserByUsername(loginUser.getUsername()), UserDTO.class), "token", token)), "登陆成功!");
-        } else {
-            throw new UserException(ResultCodeEnum.DATABASE_UPDATE_ERROR, "更新用户登陆状态失败!");
-        }
+        // 登陆成功并成功更新登陆状态后将用户信息存入redis
+        redisUtil.setObjectWithExpire("user:" + uid, loginUser.getUser(), TimeUnit.SECONDS);
+        // 返回token和用户信息给前端
+        return ResultData.success(
+                new HashMap<String, Object>(
+                        Map.of("user", BeanUtil.copyProperties(getUserByUid(uid), UserDTO.class), "token", token)),
+                "登陆成功!");
 
     }
 
@@ -131,41 +128,79 @@ public class UserServiceImpl extends ServiceImpl<UserMapper, User> implements Us
         return userMapper.selectOne(new LambdaQueryWrapper<User>().eq(User::getUsername, username));
     }
 
-//    /**
-//     * 通过用户uid更新用户信息
-//     *
-//     * @param uid 用户实体
-//     */
-//    @Transactional
-//    @Override
-//    public int updateUserById(Long uid) {
-//        return userMapper.update(new LambdaUpdateWrapper<User>().eq(User::getUid, uid).set(User::getIsLogin, true));
-//    }
+    /**
+     * 通过用户ID获取用户信息
+     *
+     * @param uid 用户ID
+     * @return user User实体
+     */
+    @Override
+    public User getUserByUid(Long uid) {
+        return userMapper.selectOne(new LambdaQueryWrapper<User>().eq(User::getUid, uid));
+    }
 
+    /**
+     * 更新用户信息
+     *
+     * @param user User实体
+     * @return int
+     */
     @Transactional
     @Override
     public int updateUserById(User user) {
         return userMapper.updateById(user);
     }
 
+    /**
+     * 用户登出
+     * 
+     * @param authorization token
+     * @return ResultData对象
+     */
+    @Transactional
     @Override
     public ResultData<String> logout(String authorization) {
-        //从SecurityContextHolder获取用户信息
+        // 从SecurityContextHolder获取用户信息
         User loginUser = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
-        if (Objects.nonNull(loginUser)) {
-            //从请求头获取Authorization并提取token
+        // 判断是否存在用户信息并且token不为空
+        if (Objects.nonNull(loginUser) && Objects.nonNull(authorization)) {
+            // 从请求头中获取token
             String token = authorization.substring(7);
-            //从jwt中获取jti
-            Object jti = jwtUtil.getClaimFromToken(token, "jti");
-            //删除redis存储的用户信息和token
+            // 从jwt中获取jti
+            String jti = jwtUtil.getClaimFromToken(token, "jti");
+            // 从redis中删除用户信息
             if (redisUtil.delete("user:" + loginUser.getUid(), "token:user:" + loginUser.getUid())) {
-                //把已删除的token的jti存入黑名单,防止token过期前有人继续用旧token
+                // 将token加入黑名单
                 redisUtil.setWithExpire("blacklist:user:" + loginUser.getUid() + ":" + jti, jti, TimeUnit.SECONDS);
                 return ResultData.success("用户登出成功");
             }
             return ResultData.fail(ResultCodeEnum.INTERNAL_SERVER_ERROR, "用户已登出");
         }
-        return ResultData.fail(ResultCodeEnum.USER_NOT_EXIST, "用户未登录");
+        return ResultData.fail(ResultCodeEnum.USER_NOT_EXIST, "用户不存在或用户未登录");
     }
 
+    /**
+     * 获取用户信息
+     *
+     * @param authorization token
+     * @return ResultData对象
+     */
+    @Override
+    public ResultData<User> getUserInfo(String authorization) {
+        // 从请求头获取Authorization并提取token
+        String token = authorization.substring(7);
+        try {
+            // 从jwt中获取uid
+            Long uid = Long.parseLong(jwtUtil.getClaimFromToken(token, "uid"));
+            // 通过uid获取用户信息
+            if (Objects.nonNull(getUserByUid(uid))) {
+                return ResultData.success(getUserByUid(uid), "获取用户信息成功");
+            } else {
+                return ResultData.fail(ResultCodeEnum.USER_NOT_EXIST, "用户不存在");
+            }
+
+        } catch (NumberFormatException e) {
+            throw new IllegalArgumentException("字符串无法转为 Long");
+        }
+    }
 }
